Revisi response get pet dan adopted pet tidak boleh delete dan update

+++ usecase/pet.go
@@ -16,9 +16,14 @@ func GetAvailablePets() (response []payload.GetPet, err error) {
 
 	for _, value := range pets {
 		response = append(response, payload.GetPet{
+			ID:        value.ID,
 			Deskripsi: value.Deskripsi,
 			Status:    value.Status,
 			Category:  value.PetCategory.Name,
+			OwnerID:   value.UserID,
+			Owner:     value.User.Name,
+			Handphone: value.User.UserDetail.Handphone,
+			Alamat:    value.User.UserDetail.Alamat,
 		})
 	}
 	return response, nil
@@ -31,9 +36,14 @@ func GetPet(id uint) (response payload.GetPet, err error) {
 	}
 
 	response = payload.GetPet{
+		ID:        pet.ID,
 		Deskripsi: pet.Deskripsi,
 		Status:    pet.Status,
 		Category:  pet.PetCategory.Name,
+		OwnerID:   pet.UserID,
+		Owner:     pet.User.Name,
+		Handphone: pet.User.UserDetail.Handphone,
+		Alamat:    pet.User.UserDetail.Alamat,
 	}
 	return response, nil
 }
@@ -45,21 +55,27 @@ func GetPets() (response []payload.GetPet, err error) {
 	}
 	for _, value := range pets {
 		response = append(response, payload.GetPet{
+			ID:        value.ID,
 			Deskripsi: value.Deskripsi,
 			Status:    value.Status,
 			Category:  value.PetCategory.Name,
+			OwnerID:   value.UserID,
+			Owner:     value.User.Name,
+			Handphone: value.User.UserDetail.Handphone,
+			Alamat:    value.User.UserDetail.Alamat,
 		})
 	}
 	return
 }
 
-func GetDonates(id uint) (response []payload.GetPet, err error) {
+func GetDonates(id uint) (response []payload.GetDonateList, err error) {
 	user, err := database.GetDonateList(id)
 	if err != nil {
 		return
 	}
 	for _, value := range user.Pet {
-		response = append(response, payload.GetPet{
+		response = append(response, payload.GetDonateList{
+			ID:        value.ID,
 			Deskripsi: value.Deskripsi,
 			Status:    value.Status,
 			Category:  value.PetCategory.Name,
@@ -68,15 +84,16 @@ func GetDonates(id uint) (response []payload.GetPet, err error) {
 	return
 }
 
-func GetAdoptions(id uint) (response []payload.GetAdoptList, err error) {
+func GetAdoptions(id uint) (response []payload.GetPet, err error) {
 	user, err := database.GetAdoptList(id)
 	if err != nil {
 		return
 	}
 	for _, value := range user.Adoption {
-		response = append(response, payload.GetAdoptList{
+		response = append(response, payload.GetPet{
 			Deskripsi: value.Pet.Deskripsi,
 			Category:  value.Pet.PetCategory.Name,
+			OwnerID:   value.UserID,
 			Owner:     value.Pet.User.Name,
 			Handphone: value.Pet.User.UserDetail.Handphone,
 			Alamat:    value.Pet.User.UserDetail.Alamat,
@@ -93,12 +110,15 @@ func UpdatePet(req *payload.UpdatePet, id uint, petId uint) error {
 	if pet.UserID != id {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Don't have permission")
 	}
+	if pet.Status == "adopted" {
+		return echo.NewHTTPError(http.StatusBadRequest, "Can't update adopted pet")
+	}
 	if err := database.UpdatePet(req, petId); err != nil {
 		return err
 	}
 	return nil
 }
-func UpdatePetStatus(id uint, petId uint) error {
+func UpdateAvailableStatus(id uint, petId uint) error {
 	pet, err := database.GetPet(petId)
 	if err != nil {
 		return err
@@ -106,7 +126,7 @@ func UpdatePetStatus(id uint, petId uint) error {
 	if pet.UserID != id {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Don't have permission")
 	}
-	if err := database.UpdatePetStatus(pet.ID, &pet); err != nil {
+	if err := database.UpdateAvailableStatus(pet.ID); err != nil {
 		return err
 	}
 	return nil
@@ -126,7 +146,7 @@ func DonatePet(req *payload.CreatePet, id uint) error {
 	return nil
 }
 
-func AdoptPet(id uint, petId uint) (response payload.GetAdoptList, err error) {
+func AdoptPet(id uint, petId uint) (response payload.GetPet, err error) {
 	profil, err := database.GetProfil(id)
 	if err != nil {
 		return
@@ -141,15 +161,16 @@ func AdoptPet(id uint, petId uint) (response payload.GetAdoptList, err error) {
 	if pet.UserID == id {
 		return response, echo.NewHTTPError(http.StatusNotAcceptable, "Owner can't adopt the pet")
 	}
-	if err := database.AdoptedStatus(petId); err != nil {
+	if err := database.UpdateAdoptedStatus(petId); err != nil {
 		return response, err
 	}
 	if err := database.AdoptPet(id, petId); err != nil {
 		return response, err
 	}
-	response = payload.GetAdoptList{
+	response = payload.GetPet{
 		Deskripsi: pet.Deskripsi,
 		Category:  pet.PetCategory.Name,
+		OwnerID:   pet.UserID,
 		Owner:     pet.User.Name,
 		Handphone: pet.User.UserDetail.Handphone,
 		Alamat:    pet.User.UserDetail.Alamat,
@@ -164,6 +185,9 @@ func DeletePet(id uint, petId uint) error {
 	}
 	if pet.UserID != id {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Don't have permission")
+	}
+	if pet.Status == "adopted" {
+		return echo.NewHTTPError(http.StatusBadRequest, "Can't delete adopted pet")
 	}
 	if err := database.DeletePet(petId); err != nil {
 		return err
